+++ src/stories/Animations.stories.ts
@@ -1,5 +1,5 @@
-import { Actor, Loader, Util } from '../engine';
-import { ImageSource, SpriteSheet, Animation, AnimationStrategy } from '../engine/Graphics';
+import { Actor, Color, Loader, Util } from '../engine';
+import { ImageSource, SpriteSheet, Animation, AnimationStrategy, Rectangle } from '../engine/Graphics';
 import { enumToControlSelectLabels, enumToControlSelectOptions, withEngine } from './utils';
 
 import animationSprite from './assets/animation.png';
@@ -10,8 +10,12 @@ export default {
 
 export const multipleFrames: Story = withEngine(async (game, { strategy }) => {
   const playerTexture = new ImageSource(animationSprite);
-
-  const player = new Actor({ x: game.currentScene.camera.x, y: game.currentScene.camera.y, width: 100, height: 30 });
+  const player = new Actor({
+    x: game.screen.center.x,
+    y: game.screen.center.y,
+    width: 100,
+    height: 30
+  });
   player.anchor.setTo(0.5, 0.5);
   const spritesheet = SpriteSheet.fromImageSource({
     image: playerTexture,
@@ -47,4 +51,90 @@ multipleFrames.argTypes = {
 };
 multipleFrames.args = {
   strategy: AnimationStrategy.End
+};
+
+export const frameDuration: Story = withEngine(async (game, { duration }) => {
+  const size = 100;
+  const colors = [Color.Red, Color.Orange, Color.Yellow, Color.Green, Color.Cyan, Color.Blue, Color.Violet];
+  const player = new Actor({
+    x: game.screen.center.x,
+    y: game.screen.center.y,
+    width: size,
+    height: size
+  });
+  const animation = new Animation({
+    frameDuration: duration,
+    frames: colors.map((c) => ({
+      graphic: new Rectangle({
+        width: size,
+        height: size,
+        color: c
+      })
+    }))
+  });
+  player.graphics.add(animation);
+  game.currentScene.add(player);
+  await game.start();
+});
+frameDuration.story = {
+  parameters: {
+    notes: 'Should display 7 frames with a delay of 1s between them'
+  }
+};
+frameDuration.argTypes = {
+  duration: {
+    name: 'Frame Duration',
+    control: {
+      type: 'range',
+      min: 100,
+      max: 2000,
+      step: 100
+    }
+  }
+};
+frameDuration.args = {
+  duration: 1000
+};
+
+export const totalDuration: Story = withEngine(async (game, { duration }) => {
+  const size = 100;
+  const colors = [Color.Red, Color.Orange, Color.Yellow, Color.Green, Color.Cyan, Color.Blue, Color.Violet];
+  const player = new Actor({
+    x: game.screen.center.x,
+    y: game.screen.center.y,
+    width: size,
+    height: size
+  });
+  const animation = new Animation({
+    totalDuration: duration,
+    frames: colors.map((c) => ({
+      graphic: new Rectangle({
+        width: size,
+        height: size,
+        color: c
+      })
+    }))
+  });
+  player.graphics.add(animation);
+  game.currentScene.add(player);
+  await game.start();
+});
+totalDuration.story = {
+  parameters: {
+    notes: 'Should display 7 frames in 1s'
+  }
+};
+totalDuration.argTypes = {
+  duration: {
+    name: 'Total Duration',
+    control: {
+      type: 'range',
+      min: 100,
+      max: 2000,
+      step: 100
+    }
+  }
+};
+totalDuration.args = {
+  duration: 1000
 };